[Agenda] enhances tool
