--- src/plugin/agenda/Resources/modules/tools/agenda/modals/parameters/store/actions.js
@@ -1,2 +1,0 @@
-
-export const actions = {}
