--- src/plugin/agenda/Resources/modules/tools/agenda/modals/parameters/store/index.js
@@ -1,10 +1,0 @@
-
-import {actions} from '#/plugin/agenda/tools/agenda/modals/parameters/store/actions'
-import {reducer} from '#/plugin/agenda/tools/agenda/modals/parameters/store/reducer'
-import {selectors} from '#/plugin/agenda/tools/agenda/modals/parameters/store/selectors'
-
-export {
-  actions,
-  reducer,
-  selectors
-}
