--- src/plugin/agenda/Resources/modules/tools/agenda/modals/parameters/store/selectors.js
@@ -1,6 +1,0 @@
-
-const STORE_NAME = 'agendaParameters'
-
-export const selectors = {
-  STORE_NAME
-}
